Rename the files validate*  -> envelope*.
